requirements.txt
anndata==0.12.6
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.1
argcomplete==3.6.3
argh==0.31.3
array-api-compat==1.12.0
astor==0.8.1
audioop-lts==0.2.2
bcbio-gff==0.7.1
beautifulsoup4==4.14.3
biopandas==0.5.1
biopython==1.86
bioregistry==0.13.10
biothings_client==0.4.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
curies==0.12.5
cycler==0.12.1
Deprecated==1.3.1
dill==0.4.0
dnaio==1.2.4
dnspython==2.8.0
donfig==0.8.1.post1
einops==0.8.1
email-validator==2.3.0
entrezpy==2.1.3
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
future==1.0.0
gast==0.7.0
gffutils==0.13
google-crc32c==1.7.1
graphtools==2.1.0
h11==0.16.0
h5py==3.15.1
harmonypy==0.0.10
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.34.4
idna==3.10
isal==1.8.0
ivy==1.0.0.5
joblib==1.5.2
kiwisolver==1.4.9
legacy-api-wrap==1.5
looseversion==1.1.2
lxml==6.0.1
matplotlib==3.10.6
mmtf-python==1.1.3
modlamp==4.3.2
mofax==0.3.7
more-click==0.1.3
msgpack==1.1.2
mygene==3.2.2
mysql-connector-python==9.5.0
myvariant==1.0.0
narwhals==2.13.0
natsort==8.4.0
networkx==3.5
nltk==3.9.1
nose==1.3.7
numcodecs==0.16.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parasail==1.3.4
patsy==1.0.2
pdb-tools==2.5.0
phate==2.0.0
pillow==11.3.0
plotly==6.5.0
propy==1.0.0
psutil==7.1.3
PubChemPy==1.0.5
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyfaidx==0.9.0.3
pyfasta==0.5.2
PyGSP==0.6.1
pyparsing==3.2.3
pysradb==2.5.1
pystow==0.7.12
pyTelegramBotAPI==4.29.1
pyteomics==4.7.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
rdkit==2025.3.6
regex==2025.9.1
requests==2.32.5
ruff==0.14.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
screed==1.1.3
seaborn==0.13.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
statsmodels==0.14.5
tangram==0.7.0
tasklogger==1.2.0
telebot==0.0.5
threadpoolctl==3.6.0
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
wrapt==2.0.1
xmltodict==1.0.2
xopen==2.0.2
zarr==3.1.5
zlib-ng==1.0.0
